requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
shapely>=2.0.0
pyshp>=2.3.0
rtree>=1.0.0
